requirements.txt
anyio==3.6.2
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.1
bcrypt==4.1.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
convertdate==2.4.0
cryptography==39.0.1
cx-Oracle==8.3.0
Django==4.1.5
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.0
fastapi==0.90.1
fastapi-pagination==0.11.4
greenlet==2.0.1
h11==0.14.0
hijri-converter==2.2.4
holidays==0.21
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
jwt==1.3.1
korean-lunar-calendar==0.3.1
loguru==0.6.0
mysqlclient==2.2.1
numpy==1.24.2
openpyxl==3.1.2
oracledb==1.2.2
packaging==23.0
pandas==1.5.3
passlib==1.7.4
pluggy==1.0.0
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
PyJWT==2.8.0
PyMeeus==0.5.12
PyMySQL==1.0.2
pytest==7.2.1
python-dateutil==2.8.2
python-decouple==3.7
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.7.1
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
SQLAlchemy-Utils==0.39.0
sqlparse==0.4.3
starlette==0.23.1
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==2.0.2
uvicorn==0.20.0
win32-setctime==1.1.0